UFLA: Collect interface equalities from theory solvers and switch to MBTC
Instead of considering all pairs of interface variables in the
preprocessing (and adding corresponding clauses) we do this lazily.
After SAT solver found a satisfying assignment, we check if theory
solvers believe some of the interface variables should be equal (if they
are not equal, the models can be unified). If there are such variables
we need to make sure this is consistent with the models found by other
theory solvers.

In Egraph, we simply check equivalence classes for interface variables.
In LASolver, we group the variables by their current Delta assignent in
the Simplex.