UFLA: Fix collection of new interface equalities
In order to deal with delta computation, we introduce new interface
equalities after everything else is settled.
However, the detection of this end state was incorrect.
It might have happened that LASolver had thought new equalities are
going to be introduced, but these were already known from Egraph, so
nothing got introduced and this final check for delta has been skipped,
resulting in an incorrect model (due to delta) once again.

The detection of the final state has been fixed as now we pass in known
equalities from the UFLATHandler, and not track known equalities
internally in LASolver.