Counting: handle correctly optimised variables
In case all boolean atoms corresponding to a term that should be counted
are optimised away, the code in principle correctly adds phony boolean
atoms to the counting line.  However, the code incorrectly searches for
the boolean atoms based on the missing term.  This results in an
assertion failure (access to a nonexisting element in a map).  This
commit fixes the issue by simply adding a check that the term exists
before accessing it.