Theory: Improve substitution-based rewriting
The substitution rewriting system is based on identifying top-level
equalities, substituting them in the original formula, and asserting the
equalities in the top level separately.

The current system is somewhat unreliable in computing the transitive
closure of substitutions.  This commit uses a union-find-style data
structure to compute the transitive closure.  The idea is taken from our
egraph algorithm, but does not compute congurences (for now?).

The system is separated from Theory and Logic and moved to a separate
class (hierarchy) in simplifiers.