Parser: correct deletion on top-level lets
In current implementation let terms' bindings' terms are deleted inside
TermNode::~TermNode.  However, ~TermNode itself cannot know whether it
is a LetTermNode.  As a result, the top-level LetTermNode's bindings are
not deleted.  The commit fixes this by adding a destructor for
LetTermNode.  The destructor is guaranteed not to recurse because
~TermNode will take care of clearing bindings vector on all
other TermNodes apart from the top-level.